split omdb into lib/bin, make dup, use it in nexus tests
