+++ dev-tools/omdb/src/crucible_pantry.rs
