+++ dev-tools/omdb/src/db/blueprints.rs
