+++ dev-tools/omdb/src/db/whatis.rs
