+++ dev-tools/omdb/src/mgs/sensors.rs
